servers/python-flask/requirements.txt
cachetools==3.1.1
certifi==2019.6.16
chardet==3.0.4
Click==7.0
Flask==1.1.1
idna==2.8
itsdangerous==1.1.0
Jinja2==2.10.1
MarkupSafe==1.1.1
numpy==1.16.4
oauthlib==3.0.2
pandas==0.24.2
pandas-gbq==0.10.0
protobuf==3.8.0
pyasn1==0.4.5
pyasn1-modules==0.2.5
pydata-google-auth==0.1.3
python-dateutil==2.8.0
python-dotenv==0.10.3
pytz==2019.1
requests==2.22.0
requests-oauthlib==1.2.0
rsa==4.0
six==1.12.0
urllib3==1.25.3
Werkzeug==0.15.4
